# Conflicts:
#	FastFlat/Areas/Identity/Pages/Account/Login.cshtml.cs
#	FastFlat/Areas/Identity/Pages/Account/LoginWithRecoveryCode.cshtml.cs
#	FastFlat/Areas/Identity/Pages/Account/Manage/Email.cshtml.cs
#	FastFlat/DAL/DBInit.cs
#	FastFlat/Migrations/RentalDbContextModelSnapshot.cs
#	FastFlat/Models/LandlordModel.cs
#	FastFlat/Models/ListningModel.cs
#	FastFlat/Models/UserModel.cs
#	FastFlat/Program.cs
#	FastFlat/RentalDatabase.db
#	FastFlat/Views/Explorer/ViewListing.cshtml

+++ FastFlat/Models/ListningModel.cs
@@ -21,7 +21,7 @@ namespace FastFlat.Models
         public int? SquareMeter { get; set; }
         public float? Rating { get; set; }
         public string? ListningAddress { get; set; } = string.Empty;
-        public double ListningPrice { get; set; }
+        public decimal ListningPrice { get; set; }
 
         //public virtual CityModel City { get; set; }//legger til city
         public virtual LocationModel? Location { get; set; } = default!;
